+++ CHANGELOG.rst
@@ -25,10 +25,12 @@ Feature:
 - Delete Damange Object will be delete no character container's character
 - Delete Damange Object will be delete broken map spawner
 - Edit Instance to Pals only
+- Copy Player can be load the Player's save file from the backup Level.sav folder
 
 Fix:
 
 - CopyPlayer on the same save file will be lost the working / base state, and put to PalStorage
+- CopyPlayer multiple times on same target UUID will be no pals on character
 - Copy Instance will be check for empty slots
 
 v0.7.2